+++ mrredis/conn_adapter.go
@@ -13,6 +13,9 @@ import (
 
 const (
 	connectionName = "Redis"
+
+	defaultReadTimeout  = 5 * time.Second
+	defaultWriteTimeout = 5 * time.Second
 )
 
 type (
@@ -23,10 +26,12 @@ type (
 
 	// Options - опции для создания соединения для ConnAdapter.
 	Options struct {
-		Host        string
-		Port        string
-		Password    string
-		ConnTimeout time.Duration
+		DSN          string // если указано, то Host, Port не используются, но Password более приоритетен если явно указан
+		Host         string
+		Port         string
+		Password     string
+		ReadTimeout  time.Duration
+		WriteTimeout time.Duration
 	}
 )
 
@@ -35,23 +40,55 @@ func New() *ConnAdapter {
 	return &ConnAdapter{}
 }
 
-// Connect - comment method.
+// Connect - создаёт соединение с указанными опциями.
 func (c *ConnAdapter) Connect(_ context.Context, opts Options) error {
 	if c.conn != nil {
 		return mrcore.ErrStorageConnectionIsAlreadyCreated.New(connectionName)
 	}
 
+	var (
+		addr string
+		db   int
+	)
+
+	if opts.ReadTimeout == 0 {
+		opts.ReadTimeout = defaultReadTimeout
+	}
+
+	if opts.WriteTimeout == 0 {
+		opts.WriteTimeout = defaultWriteTimeout
+	}
+
+	if opts.DSN != "" {
+		parsedOpts, err := redis.ParseURL(opts.DSN)
+		if err != nil {
+			return fmt.Errorf("error parsing redis DSN %s: %w", opts.DSN, err)
+		}
+
+		addr = parsedOpts.Addr
+		db = parsedOpts.DB
+
+		if opts.Password == "" {
+			opts.Password = parsedOpts.Password
+		}
+	} else {
+		addr = fmt.Sprintf("%s:%s", opts.Host, opts.Port)
+	}
+
 	c.conn = redis.NewClient(
 		&redis.Options{
-			Addr:     fmt.Sprintf("%s:%s", opts.Host, opts.Port),
-			Password: opts.Password,
+			Addr:         addr,
+			Password:     opts.Password,
+			DB:           db,
+			ReadTimeout:  opts.ReadTimeout,
+			WriteTimeout: opts.WriteTimeout,
 		},
 	)
 
 	return nil
 }
 
-// Ping - comment method.
+// Ping - проверяет работоспособность соединения.
 func (c *ConnAdapter) Ping(ctx context.Context) error {
 	if c.conn == nil {
 		return mrcore.ErrStorageConnectionIsNotOpened.New(connectionName)
@@ -69,7 +106,7 @@ func (c *ConnAdapter) Cli() redis.UniversalClient {
 	return c.conn
 }
 
-// Close - comment method.
+// Close - закрывает текущее соединение.
 func (c *ConnAdapter) Close() error {
 	if c.conn == nil {
 		return mrcore.ErrStorageConnectionIsNotOpened.New(connectionName)